Minor updates to the loot tables and a few other bits.
Updated the location_loot to categorise by cardinal, but I'm not sure how I feel about it yet. Amended the get_loot etc to match the new format. Annoyingly the standard loot table and the location loot are no longer depth equivalent but it's worked out okay I think. Updated item dropping to record the new location of the item dropped, though only on the item, not the location obj. That's next. Added the loop pref to the settings json so it doesn't ask every time if True.