Update to the new item management.
Included the updated 'item_management.py' but am not going to be using it. item_management_2 will be the basis for the new version, but it's not edited yet. Functionally, it's a transition to an id-based system, instead of purely name-string based. Will allow for duplicate items, and hopefully easier tracking as the management in-game won't be reliant on text alone, hopefully. Will need a lot of small changes in the main script but it should be doable.