Added commands to TUI, fixed drop command, fixed colours
Minor TUI update, updated the Drop and made it actually print the updated inventory. switch_the is duplicated in the main script and in misc_utilities, will switch it later but am thinking on how best to do it, given how now assign_colour works with switch_the.

Braindead today, no idea if any of this is going to stick.  But it seems to work so far.