Major colour and TUI update
Updated the game (brutally but functionally) to run in the TUI. Also routed all colours to a standalone colour script. So now the entire game runs in the TUI setup, not just the first 17 seconds. Things still need cleaning, but the base functions all work. All prints are routed through tui_update.