minor inventory progress, saving before I remove children from inv
I need to remove the children from the open inventory, will instead mark presence of known children with <item>*. Will make more sense and make container management far clearer (as currently items inside of containers are identical to outside of containers). So this is just an interim commit with a couple of minor fixes.