Container children fixed
Seems to be working well now. Adjusted some print statements to be a bit nicer, but mostly it's just about getting the containers to work properly. Haven't tested recursion but single level parenting works as expected now. The item* works nicely. Will add a note about it to the help text.