input to options comparison func, check name func.
New item membrane for interfacing str input to objects, matching case etc. Not widely implemented but local tests are good. Might be a good mainstay going forward, or at least the concept of one. Also a simple utility for checking the "*", " x" names and outputting the relevant data. Much more straightforward but good to have in a utility form explicitly.