Rearranged a heap of functions.
Attempted to add a logger (paused again because too tired to think), moved a bunch of functions around, centralised tui_enabled to a separate config file, added logging everywhere (though it barely counts as logging). Not an incredible commit but committing anyway in case I break it all trying to work on it later today. Rough day.