Testing verb objects
Wrote some starting classes/dicts for verb objects. Not sure if it's a direction I want to go in, but could be interesting. Instead of relying on pure strings, give the option of verb objects (can still be printed the same way) that hold their own alt_names and define viable format options. So instead of the script having to define 'if 'drop' is typed, do xyz', it would be the domain of the 'drop object' to know what to do. Would make it more flexible to use actions across scripts, as it would be defined in the verb-instance-space, not the active script, if that makes sense.